Perf: evaluate issue paths lazily
refactors 'path' to use 'class LazyPath' instead of
PathItem[].

Issue paths are only needed when flattening errors, so
eagerly constructing them for the ParseInfo objects passed
to every parser results in a lot of unnecessary allocations.

Skipping most of these allocations and evaluating the paths
only when they are needed improves performance.