Refactor Fallback and Default methods, Add Schema Type Predicates
Using some basic type predicate functions for base schemas, object schemas, and tuple schemas, I refactored the fallback and default methods which simplifies both their types and the logic for each method.

Note that because each of these methods relied on a mutable return value typed as `any`, Typescript did not catch any logic errors when calculating default or fallback values. With the refactors, in some cases an `as` type assertion is necessary.

For some of the method types, I reorganized them such that they no longer introduce a circular dependency (even if only at the type level).