Add support for 9-digit millisecond precision to isoTimestamp #348
ISO-8601 does not specify the number of digits in the decimal fraction.

JavaScript `new Date().toISOString()` which is compliant to ISO-8601, has a precision of three digits (miliseconds precision).
Java `Instant.now().toString()` which is compliant to ISO-8601, has a precision of nine digits (nanoseconds precision).

This change will allow timestamps with nanoseconds precision.