feat: disallow additional props in intersections
Sets `unevaluatedProperties: false` in intersections which have
`additionalProperties` consistently set to `false`.

This will enforce that to validate against the JSON schema, your input
must have only properties from each of the intersecting schemas.

If any of the sub-schemas have `additionalProperties` set to a non-false
value, `unevaluatedProperties` will not be set.