Fix stringification of objects with null prototype
`Object.getPrototypeOf` can return `null`
if the object passed to it has a `null` prototype.
If that is the case,
we can not access any properties
on the `null` prototype.

Use optional chaining
after getting the object's prototype
to attempt to get the name
of the prototype's constructor.
Then check to see if `name` is a string
before checking to see if the constructor's name
is something we want to use
in place of the value
in the stringified result.

This fixes fabian-hiller/valibot#1114.