Add `valipass` to Ecosystem page
Added a link to [`valipass`](https://github.com/Saeris/valipass) to the Ecosystem page.

This library consists of a handful of validation actions and a `password` schema with some sensible password requirement defaults. It is functionally a Valibot version of [`yup-password`](https://github.com/knicola/yup-password) and is meant to be used in client-side form validation scenarios.